item production price set to decimal data type

+++ src/Resourcerer.DataAccess/Entities/Item.cs
@@ -12,7 +12,7 @@ public class Item : AppDbEntity
     }
 
     public string? Name { get; set; }
-    public double ProductionPrice { get; set; }
+    public decimal ProductionPrice { get; set; }
     public double ProductionTimeSeconds { get; set; }
     public double? ExpirationTimeSeconds { get; set; }
