+++ src/Resourcerer.DataAccess/Entities/JsonEntities/Instance/Events/Order/InstanceOrderCancelledEvent.cs
@@ -5,5 +5,5 @@ namespace Resourcerer.DataAccess.Entities;
 public class InstanceOrderCancelledEvent : AppDbJsonField
 {
     public string? Reason { get; set; }
-    public double RefundedAmount { get; set; }
+    public decimal RefundedAmount { get; set; }
 }
