+++ src/Resourcerer.Dtos/V1/Instances/V1InstanceInfo.cs
@@ -7,7 +7,7 @@ public class V1InstanceInfo : IDto
     public double QuantityLeft { get; set; }
     public V1DiscardInfo[]? Discards { get; set; }
     public DateTime? ExpiryDate { get; set; }
-    public double PurchaseCost { get; set; }
-    public double SellProfit { get; set; }
-    public float SellCancellationsPenaltyDifference { get; set; }
+    public decimal PurchaseCost { get; set; }
+    public decimal SellProfit { get; set; }
+    public decimal SellCancellationsPenaltyDifference { get; set; }
 }
